# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ai/cash-flow-forecast": {
      "get": {
        "operationId": "AiController_getCashFlowForecast",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Get AI-powered 30/60/90 day cash flow forecast",
        "tags": [
          "AI"
        ]
      }
    },
    "/ai/insights": {
      "get": {
        "operationId": "AiController_getInsights",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Get AI-generated business insights feed",
        "tags": [
          "AI"
        ]
      }
    },
    "/ai/anomalies": {
      "get": {
        "operationId": "AiController_getAnomalies",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Detect anomalies in recent transactions",
        "tags": [
          "AI"
        ]
      }
    },
    "/ai/categorize-transaction": {
      "post": {
        "operationId": "AiController_categorizeTransaction",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "description",
                  "amount"
                ],
                "properties": {
                  "description": {
                    "type": "string",
                    "example": "UPI/AMAZON RETAIL/STATIONERY"
                  },
                  "amount": {
                    "type": "number",
                    "example": 1499
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "AI-suggest account category for a bank transaction",
        "tags": [
          "AI"
        ]
      }
    },
    "/ai/reconciliation-suggest": {
      "post": {
        "operationId": "AiController_suggestReconciliation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "transactionId"
                ],
                "properties": {
                  "transactionId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "8b6b0c1e-2f5a-4b3c-9d8e-1a2b3c4d5e6f"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Suggest matching journal entries for bank reconciliation",
        "tags": [
          "AI"
        ]
      }
    },
    "/ai/settings": {
      "get": {
        "operationId": "AiController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Get AI feature toggle states",
        "tags": [
          "AI"
        ]
      },
      "patch": {
        "operationId": "AiController_updateSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cashFlowForecast": {
                    "type": "boolean",
                    "example": true
                  },
                  "anomalyDetection": {
                    "type": "boolean",
                    "example": true
                  },
                  "insightGeneration": {
                    "type": "boolean",
                    "example": true
                  },
                  "transactionCategorization": {
                    "type": "boolean",
                    "example": true
                  },
                  "reconciliationAssist": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Enable/disable individual AI features",
        "tags": [
          "AI"
        ]
      }
    },
    "/ai/stored-insights": {
      "get": {
        "operationId": "AiController_getStoredInsights",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Get persisted AI insights for dashboard widget",
        "tags": [
          "AI"
        ]
      }
    },
    "/ai/dismiss-insight": {
      "post": {
        "operationId": "AiController_dismissInsight",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "insightId"
                ],
                "properties": {
                  "insightId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "8b6b0c1e-2f5a-4b3c-9d8e-1a2b3c4d5e6f"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Dismiss an AI insight",
        "tags": [
          "AI"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kontafy API",
    "description": "REST API for Kontafy — cloud-native accounting for Indian businesses. All paths are versioned: prefix every request with `/v1`. Authenticated routes expect `Authorization: Bearer <access_token>` and most multi-tenant routes also expect `X-Org-Id: <organization id>`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.kontafy.com/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:5002/v1",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token returned by POST /v1/auth/login"
      },
      "org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Org-Id",
        "description": "Organization id for multi-tenant routes"
      }
    },
...
  }
}